separate list splitting into utils
reoganize utils
