+++ common/src/main/java/com/wynntils/core/webapi/request/RequestBuilder.java
@@ -9,8 +9,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.wynntils.core.webapi.WebReader;
-import com.wynntils.utils.MD5Verification;
 import com.wynntils.utils.Utils;
+import com.wynntils.utils.objects.MD5Verification;
 import com.wynntils.utils.objects.ThrowingBiPredicate;
 import java.io.File;
 import java.io.IOException;
