+++ common/src/main/java/com/wynntils/features/SoulPointTimerFeature.java
@@ -9,8 +9,8 @@ import com.wynntils.core.features.GameplayImpact;
 import com.wynntils.core.features.PerformanceImpact;
 import com.wynntils.core.features.Stability;
 import com.wynntils.mc.event.InventoryRenderEvent;
-import com.wynntils.mc.utils.DynamicTag;
 import com.wynntils.mc.utils.ItemUtils;
+import com.wynntils.mc.utils.objects.DynamicTag;
 import com.wynntils.utils.Utils;
 import com.wynntils.wc.utils.InventoryData;
 import com.wynntils.wc.utils.ItemMatchers;
