+++ common/src/main/java/com/wynntils/mc/utils/objects/DynamicTag.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2021.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.mc.utils;
+package com.wynntils.mc.utils.objects;
 
 import java.io.DataInput;
 import java.io.DataOutput;
