+++ common/src/main/java/com/wynntils/wc/impl/WorldStateImpl.java
@@ -12,7 +12,7 @@ import com.wynntils.mc.event.PlayerInfoFooterChangedEvent;
 import com.wynntils.mc.event.PlayerTeleportEvent;
 import com.wynntils.mc.event.ResourcePackEvent;
 import com.wynntils.mc.event.ScreenOpenedEvent;
-import com.wynntils.mc.utils.StringUtils;
+import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.utils.Utils;
 import com.wynntils.wc.event.WorldStateEvent;
 import com.wynntils.wc.model.WorldState;
@@ -90,7 +90,7 @@ public class WorldStateImpl implements WorldState {
 
         if (e.getId().equals(WORLD_UUID)) {
             Component displayName = e.getDisplayName();
-            String name = StringUtils.getUnformatted(displayName);
+            String name = ComponentUtils.getUnformatted(displayName);
             Matcher m = WORLD_NAME.matcher(name);
             if (m.find()) {
                 String worldName = m.group(1);
